feat: 예약 조회 API DB 연결

+++ src/main/java/roomescape/controller/dto/ReservationCreateRequest.java
@@ -7,7 +7,7 @@ import java.time.LocalTime;
 
 public record ReservationCreateRequest(String name, LocalDate date, LocalTime time) {
 
-    public Reservation toReservation(int id) {
+    public Reservation toReservation(Long id) {
         return new Reservation(id, name, date, time);
     }
 }
